sync

+++ modules/gapi/include/opencv2/gapi/own/mat.hpp
@@ -17,6 +17,7 @@
 #include <memory>                   //std::shared_ptr
 #include <cstring>                  //std::memcpy
 #include <numeric>                  //std::accumulate
+#include <vector>
 #include <opencv2/gapi/util/throw.hpp>
 
 namespace cv { namespace gapi { namespace own {
@@ -297,10 +298,9 @@ namespace cv { namespace gapi { namespace own {
          */
         size_t total() const
         {
-            return static_cast<std::size_t>
-                (dims.empty()
-                 ? (rows * cols)
-                 : std::accumulate(dims.begin(), dims.end(), 1, std::multiplies<int>()));
+            return dims.empty()
+                 ? (static_cast<std::size_t>(rows) * cols)
+                 : std::accumulate(dims.begin(), dims.end(), static_cast<std::size_t>(1), std::multiplies<size_t>());
         }
 
         /** @overload
